# fix moe_align1 kernel performance issue in prefill stage.